requirements.txt
defusedxml==0.4.1
Django==1.10
django-allauth==0.26.1
django-braces==1.9.0
django-cors-headers==1.1.0
django-debug-panel==0.8.3
django-debug-toolbar==1.5
django-filter==0.15.2
django-grappelli==2.8.1
django-multiselectfield==0.1.3
django-oauth-toolkit==0.10.0
djangorestframework==3.4.5
geopy==1.11.0
Markdown==2.6.6
mysqlclient==1.3.7
oauthlib==1.0.3
Pillow==3.3.0
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
requests==2.11.0
requests-oauthlib==0.6.2
six==1.10.0
sqlparse==0.2.1
